Refactoring server application service to enable modularity and callbacks, as well as provide app, db, config variables outside of server.js
